Crate monstorr_lib

This crate is the library for the functionality of the Monstorr program. It contains a chaotic good mess of structures and functionality for calculating D&D 5e creature stats and producing stat blocks in various formats. It’s not currently intended as a public API, and any published functionality here is very unstable. However, some of the stuff in this crate is public API, therefore it needs documentation.

Project Status. This project is newly created and poorly tested. It works, but there may be weird issues, please report any that you find so I can fix them. While feature suggestions are also appreciated, they will probably not be addressed anytime soon. One great way to test this would be to convert the SRD monsters over to creature files, which could then be included in the monstorr-data crate.

Points of Interest

There are a few things that you may need to know in order to simply use Monstorr.

For the Monstorr command line tool usage, run monstorr help.

For the creature file syntax, see crate::creature_commands.

Development

Organization and Coding Style. I originally intended this as a creature database tool which would pull in Open5e-style JSON, verifying calculations, store them for later use, and spitting out LaTeX documents (for my own needs) as needed. Little thought was put into organization, formatting and keeping to the rusty ways of doing things. It was only later when I realized the true goal of this tool that I began to clean a few things up. I apologize for this confusion, you’re welcome to fork this code and clean it up. Testing would be most appreciated.

Modules

actions 🔒

attacks 🔒

creature 🔒

creature_commands 🔒

This crate implements the syntax for creature files.

dice 🔒

dice_expression 🔒

errors 🔒

features 🔒

interpolation 🔒

In Monstorr, interpolation is the process of substituting variables and expressions with actual values at the last minute. There are actually two types of interpolation that happen in Monstorr, one occurs during the generation of stat blocks from feature descriptions and other strings, the other occurs when including an external creature file into another (see crate::creature_commands::CreatureCommand::Include). They both work basically the same, but have slightly different inputs and outputs.

open5e_convertor 🔒

parse_position 🔒

reactions 🔒

spellcasting 🔒

stat_block 🔒

Creature stat blocks are a simple property/value mapping, with no functionality. They are intended to be output directly as JSON, or used in a template to generate content for publication.

stats 🔒

structured_text 🔒

template 🔒

tokenizer 🔒

utils 🔒

Macros

digit

ident_start

whitespace

Enums

InputFormat

ListInputFormat

OutputFormat

Constants

MONSTORR_VERSION

Functions

create_stat_block

generate_creatures_as_rust_array

get_default_working_dir 🔒

get_source_file 🔒

get_working_dir_relative_to_source_or_default 🔒

list_creatures

list_template_names

print_template

read_source 🔒

resolve_existing_dir 🔒

resolve_existing_file 🔒

resolve_file 🔒

validate_creature

write_target 🔒