Enum monstorr_lib::actions::Action

pub enum Action {

    ChangeShape(String),
    Charm(u8, u8),
    Enlarge,
    Etherealness,
    FrightfulPresence(u8, u8, bool),
    HorrifyingVisage(u8, u8),
    IllusoryAppearance(u8),
    Leadership,
    Possession(u8),
    ReadThoughts(u8),
    Reel,
    Swallow(DiceExpression, u8, u8, u8),
    Teleport(u8, Option<String>),
    Attack(String, Attack, AttackEffect, Option<CompoundAttackEffect>),
    Action(String, String, Option<AttackEffect>, Option<CompoundAttackEffect>),
    BreathWeapon(String, String, Option<AttackEffect>, Option<CompoundAttackEffect>),

}

Many of the variants in this value are attempts to mimic common actions found in the SRD, as well as actions which the Dungeon Master’s Guide says may have an effect on the challenge rating. However, since the SRD does not use consistent phrasing for all actions, you may not get the output you expect. If you want to use an action with one of the specified names, use that, even if you have to override the description to fix it later. That way, you’ll get the benefits of challenge rating calculation when that is available.

Variants

ChangeShape(String)

ChangeShape(<string>)

The specified string is a description of the form the creature can turn into, such as “a humanoid or beast that has a challenge rating no higher than its own”

Charm(u8, u8)

Charm(<integer>,<integer>)

The first number is the range of the effect, and the second is the Wisdom save DC.

Enlarge

Enlarge

Etherealness

Etherealness

FrightfulPresence(u8, u8, bool)

FrightfulPresence(<integer>,<integer>,<boolean>)

The first integer is the range of the effect, the second is the Wisdom save DC. The last boolean indicates whether re-saves are done at a disadvantage, as with the Tarrasque.

HorrifyingVisage(u8, u8)

HorrifyingVisage(<integer>,<integer>)

The first integer is the range of the effect, the second is the Wisdom save DC.

IllusoryAppearance(u8)

IllusoryAppearance(<integer>)

The integer is the investigation (Intelligence) check DC to discern the true appearance.

Leadership

Leadership

Possession(u8)

Possession(<integer>)

The integer is the Charisma save DC.

ReadThoughts(u8)

ReadThoughts(<integer>)

The integer is the range of the effect.

Reel

Reel

Swallow(DiceExpression, u8, u8, u8)

Swallow(<dice-expression-string>,<integer>,<integer>,<integer>)

The first argument is the acid damage inflected while swallowed. The second is the damage required to cause the creature to vomit. The third is the Constitution save DC the creature must roll. The last is the distance required to escape.

Teleport(u8, Option<String>)

Teleport(<integer>) Teleport(<integer>,option(<string>))

The first argument is the distance the creature can teleport. The second is the name of an optional action which can be taken prior to teleporting.

Attack(String, Attack, AttackEffect, Option<CompoundAttackEffect>)

Attack(<string>,<Attack>,<AttackEffect>) Attack(<string>,<Attack>,<AttackEffect>,option(<CompoundAttackEffect>))

This variant adds a custom attack to the creature. The first argument is the name of the attack, the other arguments describe the attack, and are used to generate the description. See crate::attacks::Attack, crate::attacks::AttackEffect, and crate::attacks::CompoundAttackEffect.

Action(String, String, Option<AttackEffect>, Option<CompoundAttackEffect>)

Action(<string>,<string>,option(<AttackEffect>)) Action(<string>,<string>,option(<AttackEffect>),option(<CompoundAttackEffect>))

This variant adds a custom action that isn’t an attack, or at least doesn’t require an attack roll to hit. The first argument is the name, the second is the description, which will be interpolated. The remaining arguments describe the effects of the action, and are used to calculate damage per round. These effects aren’t used to calculate the final description.

BreathWeapon(String, String, Option<AttackEffect>, Option<CompoundAttackEffect>)

BreathWeapon(<string>,<string>,option(<AttackEffect>)) BreathWeapon(<string>,<string>,option(<AttackEffect>),option(<CompoundAttackEffect>))

Follows the same signatures as the Action variant. Breath weapon actions vary so much between creatures, with each species having their own unique powers, that these are just custom actions. But, their existence does change the challenge rating.

Implementations

impl Action

pub fn get_name(&self) -> String

pub fn get_description(&self) -> String

pub fn get_attack(&self) -> Option<Attack>

pub fn get_effect(&self) -> Option<AttackEffect>

pub fn get_compound_effect(&self) -> Option<CompoundAttackEffect>