Enum monstorr_lib::attacks::Weapon

pub enum Weapon {

    UnarmedStrike(i8),
    Club(i8),
    Dagger(i8),
    Greatclub(i8),
    Handaxe(i8),
    Javelin(i8),
    LightHammer(i8),
    Mace(i8),
    Quarterstaff(i8),
    Sickle(i8),
    Spear(i8),
    LightCrossbow(i8),
    Dart(i8),
    Shortbow(i8),
    Sling(i8),
    Battleaxe(i8),
    Flail(i8),
    Glaive(i8),
    Greataxe(i8),
    Greatsword(i8),
    Halberd(i8),
    Lance(i8),
    Longsword(i8),
    Maul(i8),
    Morningstar(i8),
    Pike(i8),
    Rapier(i8),
    Scimitar(i8),
    Shortsword(i8),
    Trident(i8),
    WarPick(i8),
    Warhammer(i8),
    Whip(i8),
    Blowgun(i8),
    HandCrossbow(i8),
    HeavyCrossbow(i8),
    Longbow(i8),
    Net(i8),

}

These values are used to represent standard weapons in attacks. Custom weapons are made by added custom attack actions. All weapons take an integer which represents a bonus (or penalty) to be added to attack and damage rolls, such as might come from a magic weapon.

Variants

UnarmedStrike(i8)

UnarmedStrike(<integer>)

Tuple Fields of UnarmedStrike

0: i8

Club(i8)

Club(<integer>)

Tuple Fields of Club

0: i8

Dagger(i8)

Dagger(<integer>)

Tuple Fields of Dagger

0: i8

Greatclub(i8)

Greatclub(<integer>)

Tuple Fields of Greatclub

0: i8

Handaxe(i8)

Handaxe(<integer>)

Tuple Fields of Handaxe

0: i8

Javelin(i8)

Javelin(<integer>)

Tuple Fields of Javelin

0: i8

LightHammer(i8)

LightHammer(<integer>)

Tuple Fields of LightHammer

0: i8

Mace(i8)

Mace(<integer>)

Tuple Fields of Mace

0: i8

Quarterstaff(i8)

Quarterstaff(<integer>)

Tuple Fields of Quarterstaff

0: i8

Sickle(i8)

Sickle(<integer>)

Tuple Fields of Sickle

0: i8

Spear(i8)

Spear(<integer>)

Tuple Fields of Spear

0: i8

LightCrossbow(i8)

LightCrossbow(<integer>)

Tuple Fields of LightCrossbow

0: i8

Dart(i8)

Dart(<integer>)

Tuple Fields of Dart

0: i8

Shortbow(i8)

Shortbow(<integer>)

Tuple Fields of Shortbow

0: i8

Sling(i8)

Sling(<integer>)

Tuple Fields of Sling

0: i8

Battleaxe(i8)

Battleaxe(<integer>)

Tuple Fields of Battleaxe

0: i8

Flail(i8)

Flail(<integer>)

Tuple Fields of Flail

0: i8

Glaive(i8)

Glaive(<integer>)

Tuple Fields of Glaive

0: i8

Greataxe(i8)

Greataxe(<integer>)

Tuple Fields of Greataxe

0: i8

Greatsword(i8)

Greatsword(<integer>)

Tuple Fields of Greatsword

0: i8

Halberd(i8)

Halberd(<integer>)

Tuple Fields of Halberd

0: i8

Lance(i8)

Lance(<integer>)

Tuple Fields of Lance

0: i8

Longsword(i8)

Longsword(<integer>)

Tuple Fields of Longsword

0: i8

Maul(i8)

Maul(<integer>)

Tuple Fields of Maul

0: i8

Morningstar(i8)

Morningstar(<integer>)

Tuple Fields of Morningstar

0: i8

Pike(i8)

Pike(<integer>)

Tuple Fields of Pike

0: i8

Rapier(i8)

Rapier(<integer>)

Tuple Fields of Rapier

0: i8

Scimitar(i8)

Scimitar(<integer>)

Tuple Fields of Scimitar

0: i8

Shortsword(i8)

Shortsword(<integer>)

Tuple Fields of Shortsword

0: i8

Trident(i8)

Trident(<integer>)

Tuple Fields of Trident

0: i8

WarPick(i8)

WarPick(<integer>)

Tuple Fields of WarPick

0: i8

Warhammer(i8)

Warhammer(<integer>)

Tuple Fields of Warhammer

0: i8

Whip(i8)

Whip(<integer>)

Tuple Fields of Whip

0: i8

Blowgun(i8)

Blowgun(<integer>)

Tuple Fields of Blowgun

0: i8

HandCrossbow(i8)

HandCrossbow(<integer>)

Tuple Fields of HandCrossbow

0: i8

HeavyCrossbow(i8)

HeavyCrossbow(<integer>)

Tuple Fields of HeavyCrossbow

0: i8

Longbow(i8)

Longbow(<integer>)

Tuple Fields of Longbow

0: i8

Net(i8)

Net(<integer>)

Tuple Fields of Net

0: i8

Implementations

impl Weapon

pub fn get_attack(&self) -> Attack

pub fn get_effect(&self, size: &CreatureSize) -> AttackEffect