Module monstorr_lib::creature_commands

This crate implements the syntax for creature files.

Rusty-Object Notation

If you just want to see the syntax, and aren’t interested in rust programming details, you can skip this section.

Creature files follow a serialization format convention called RON, or Rusty-Object Notation. This format is related to how programs written print debug information about data structures. The main reason this format is used is because I don’t have to write a special parser, as that code is already available through combination with the serde crate. RON has two features that make it great for this usage:

Structures are tagged outside the enclosing data, so there’s no need for a ‘type’ property on every object, which would make writing these creatures frustratingly wordy.

RON automatically deduces the namespace of objects based on their context. This removes the need to specify the type of an enum variant when it’s in use. This essentially turns the creature command data structure into context-sensitive functions.

Basic Syntax

RON files are encoded in utf-8. The RON notation supports several basic values. There are a few more than what is described here, but those are not used in creature commands. Some of these are called differently in the official spec.

Numbers. Numbers are written as you’d expect. They can be signed, and include decimals and scientific exponent parts (‘e+2’). Most of the types here require simple integers, but floating point values are legal in the appropriate context. The context determines the type required, and will often limit the range of the numbers due to the data type the numbers are stored in. In the documentation, these values shown as <integer> or <number> if the number is not an integer.

Strings. Strings are written using double-quotes. Conventional c-style escape sequences are allowed for certain control characters, and characters derived from unicode hex values. In the documentation, these values are shown as <string>

Booleans. Occasionally, you may find boolean (on/off) values, which are written simply as ‘true’ or ‘false’. In the documentation, these values are shown as <boolean>

List. A list value is written between brackets, with the items separated by commas ([ value, ...]). The type of values allowed in the list depends on the context. In the documentation, these values are shown as [<type>...].

Map. A mapping of keys to values. This is represented as comma-separated pairs inside brackets ({ value: value, ...}). In general the keys are strings, but they can be any type as long as they are all the same. Values must all be the same type as well. This is rarely used, but it is used at least once. In the documentation, these values are shown as <map(<key type>, <value type>)>.

Tuple Structs. A tuple struct is represented by list of plain values in parentheses, with the values separated by commas (( value, ..)). The empty tuple (()) is called the unit struct. This is different from a list, which has a variable length of values all of the same type. A tuple has a specific length and can require different value types for each member. Sometimes, the later items in the tuple can be skipped if they are optionals. In the documentation, these values are shown as (<type>,...). If a value is not required, the documentation will show all the possible signatures.

Mapped Structs. A named struct is similar to a tuple struct, but each value is associated with a named property, which is followed by a colon (( property: value, ..)). The required fields in this struct, and their value types, depend on the context they are used. In the documentation, these values are shown as ( <property>: <type>,...). If a property is not required, it will be shown with a question mark after.

Tagged Variants. Some values are differentiated by a tag because more than one value is allowed in the context. Each of these is represented by a simple identifier, usually capitalized on the first character. A tagged variant can have associated data, which is represented by following that tag either by a tuple struct or a mapped struct. In the documentation, these values are shown by the identifier.

Optionals. A special case of tagged variants is the optionals. These are represented with the simple tag None, representing a value that doesn’t exist, and a tagged tuple Some(..) representing the actual value. In many cases, if the value is None, it can also be skipped. For example, if a mapped struct has a property foo which should hold an optional string, it doesn’t have to be specified if that value would be None. In the documentation, these values are shown as optional(<type>).

The last part, the tagged variants, allow us to create what vaguely looks like a function in other languages.

Dice Expressions

Some of the arguments require dice expressions. As RON has no concept of dice, these are always passed sa strings, which must match a specific pattern. There are three types of these:

Die String. This is a reference to a single die type, in the format “d6” or “d8”. In the documentation, these values are shown with <die-string>.

Dice String. This is a reference to a die and the number of times it is rolled, in the format “1d4” or “2d10”. In the documentation, these values are shown with <dice-string>.

Dice Expression. This is a reference to a full dice expression, including bonuses, additional terms, etc, in formats like “3d6 + 10” or “2d4 - 1d10 + 2”. They can get slightly more complex than that, as each dice term could be accompanied by a multiplication factor, but you are unlikely to need that. In the documentation, these values are shown with <dice-expression-string>.

Interpolation

Some of the structures allow for strings containing interpolation variables, which allow you to generate this text based on the creature values. This is not part of the RON syntax, but something programmed specifically for Monstorr. For more information on this interpolation syntax, see crate::interpolation. For a list of the variables available in the interpolation, see crate::creature::Creature.

Creature Document Structure

For a description fo the basic creature document structure, see CreatureCreator, which is the structure deserialized from a creature file.

Structs

CreatureCreationHooks 🔒

CreatureCreator

([<CreatureCommand>...])

Enums

CreatureCommand

A CreatureCommand represents a task that is completed in the process of creating a creature to generate a stat block. Most of the commands set or add data to the resulting creature. A few commands have other effects, or remove data that has been added.

InnateSpellcastingCommand

These commands add information to a creature’s innate spellcasting feature.

SpellcastingCommand

These commands add information to a creature’s spellcasting feature.

Constants

MONSTORR_VERSION