Enum monstorr_lib::creature_commands::CreatureCommand

pub enum CreatureCommand {

    Monstorr(f32, Option<f32>),
    Include(String, HashMap<String, String>),
    IncludeStored(String),
    Source(String),
    Name(String),
    SubjectName(String),
    CapitalizeSubjectName(String),
    PossessiveName(String),
    CapitalizePossessiveName(String),
    Pronoun(String),
    SubjectPronoun(String),
    CapitalizeSubjectPronoun(String),
    PossessivePronoun(String),
    CapitalizePossessivePronoun(String),
    ObjectPronoun(String),
    ReflexivePronoun(String),
    Tiny,
    Small,
    Medium,
    Large,
    Huge,
    Gargantuan,
    Aberration,
    Beast,
    Celestial,
    Construct,
    Dragon,
    Elemental,
    Fey,
    Fiend,
    Giant,
    Humanoid,
    Monstrosity,
    Ooze,
    Plant,
    Undead,
    CreatureType(String),
    Subtype(String),
    Group(String),
    AnyAlignment,
    AnyNonGood,
    AnyNonEvil,
    AnyNonLawful,
    AnyNonChaotic,
    AnyGood,
    AnyEvil,
    AnyLawful,
    AnyChaotic,
    LawfulGood,
    NeutralGood,
    ChaoticGood,
    LawfulNeutral,
    Neutral,
    ChaoticNeutral,
    LawfulEvil,
    NeutralEvil,
    ChaoticEvil,
    Unaligned,
    Alignment(String),
    HitDie(Die),
    HitDiceCount(u8),
    HitPoints(u16),
    Armor(Armor),
    Shield,
    NoShield,
    Walk(u8),
    Swim(u8),
    Fly(u8),
    Hover,
    Burrow(u8),
    Climb(u8),
    SpeedNotes(String),
    Speed(String, u8),
    Str(u8),
    Dex(u8),
    Con(u8),
    Int(u8),
    Wis(u8),
    Cha(u8),
    Saves(Vec<Ability>),
    Skills(Vec<Skill>),
    Expertise(Vec<Skill>),
    RemoveSaves(Vec<Ability>),
    RemoveSkills(Vec<Skill>),
    ConditionImmunity(Condition),
    Vulnerability(Damage),
    AllVulnerability,
    CustomVulnerability(String),
    Resistance(Damage),
    AllResistance,
    NonmagicalResistance,
    NonSilveredResistance,
    NonAdamantineResistance,
    CustomResistance(String),
    Immunity(Damage),
    AllImmunity,
    NonmagicalImmunity,
    NonSilveredImmunity,
    NonAdamantineImmunity,
    CustomImmunity(String),
    Languages(Vec<Language>),
    UnspokenLanguages(Vec<Language>),
    Darkvision(u8),
    Blindsight(u8),
    BlindsightBlindBeyond(u8),
    Truesight(u8),
    Tremorsense(u8),
    CustomSense(String, u8),
    ExpectNoChallenge,
    ExpectChallenge(u8),
    ExpectHalfChallenge,
    ExpectQuarterChallenge,
    ExpectEighthChallenge,
    OverrideNoChallenge,
    OverrideChallenge(u8),
    OverrideHalfChallenge,
    OverrideQuarterChallenge,
    OverrideEighthChallenge,
    Multiattack(String, Multiattack),
    Weapon(Weapon, Option<CompoundAttackEffect>),
    ExpectWeaponAttack(Weapon, Attack),
    ExpectWeaponEffect(Weapon, AttackEffect),
    OverrideWeaponAttack(Weapon, Attack),
    OverrideWeaponEffect(Weapon, AttackEffect),
    OverrideWeaponDescription(Weapon, String),
    Action(Action, Option<UsageLimit>),
    OverrideActionDescription(String, String),
    RemoveWeapon(Weapon),
    MoveWeapon(Weapon, i8),
    RemoveAction(String),
    MoveAction(String, i8),
    Reaction(Reaction, Option<UsageLimit>),
    RemoveReaction(String),
    Feature(Feature, Option<UsageLimit>),
    RemoveFeature(String),
    Spellcasting(Vec<SpellcastingCommand>),
    InnateSpellcasting(Vec<InnateSpellcastingCommand>),
    LegendaryActions(u8, Vec<LegendaryAction>),
    RemoveLegendaryAction(String),
    LairActions(String, Vec<String>, String),
    RegionalEffects(String, Vec<String>, String),

}

A CreatureCommand represents a task that is completed in the process of creating a creature to generate a stat block. Most of the commands set or add data to the resulting creature. A few commands have other effects, or remove data that has been added.

While there is no strict order requirement, commands can work differently depending on what order they are added to the list. Listed statistics, like actions, are added in the order they appear in list. Other commands require the existence of other data, such as an action or other named feature, in order to work. Commands which set simple statistics, such as armor, hit points, size, etc. will override previous commands that set the same statistic.

The documentation below is automatically produced with tools which assume you are programming in Rust. If you are only interested in the commands for use in building creatures, pay no attention to the complicated signatures below the comments. The signature for use in the creature file is provided in the notes.

Variants

Monstorr(f32, Option<f32>)

Monstorr(<number>) Monstorr(<number>,Some(<number>))

Specifies the minimum version of the Monstorr creature format for which your creature is defined, and optionally, a maximum version. If the copy of Monstorr this creature is run with does not meet the requirements, an error will be reported and the creature will not be created.

The API for Monstorr is unstable. This means that commands may work differently in future versions. In general, if a command is removed, or it’s data requirements changed, your creature file will simply fail to load. However, there may be some differences in the actual functionality between certain versions of the code. This command can be used to make sure you get an error if you’re using it with an unsupported version, so you don’t accidentally get unexpected data. *

Tuple Fields of Monstorr

0: f321: Option<f32>

Include(String, HashMap<String, String>)

Include(<string>,<map(string:string)>)

This allows you to include commands from another creature file on disk. With this command, you can either take an existing creature and modify it, or add a common set of features to more than one creature.

The file is not expected to be a pure creature format. It will be interpolated in a similar way to how feature descriptions are interpolated. The second argument allows you to pass named arguments to that interpolation, letting you create more dynamic content.

The output of this interpolation is expected to be in valid Monstorr creature format.

Open5e JSON format is not supported for this because converting is a rigorous task that would require parsing the English language (an early attempt to do this for Monstorr ended with frustration). If you convert an existing creature from Open5e or the SRD, your contributions of that document to this project are welcome.

Tuple Fields of Include

0: String1: HashMap<String, String>

IncludeStored(String)

IncludeStored(<string>)

This allows you to include commands to build one of the SRD creatures stored in the application itself. This command allows you to take that existing creature and modify it with additional commands.

Tuple Fields of IncludeStored

0: String

Source(String)

Source(<string>)

If you wish to assign a source tag to the resulting creature, to give credit in the stat-block itself, you can use this command.

Tuple Fields of Source

0: String

Name(String)

Name(<string>)

Sets the name of the creature as it will appear in the title of the stat-block.

Tuple Fields of Name

0: String

SubjectName(String)

SubjectName(<string>)

When interpolating descriptions, it is often necessary to refer to the creature by name. The interpolation variable subj is used when the creature is referred to as the subject of an action. By default, the value returned is the name of the creature with a definite article (‘the goblin’). If you need to override this somehow, you can set a different value here.

Tuple Fields of SubjectName

0: String

CapitalizeSubjectName(String)

CapitalizeSubjectName(<string>)

When interpolating descriptions, it is often necessary to refer to the creature by name. The interpolation variable Subj is used when the creature is referred to as the subject of an action at the beginning of a sentence. By default, the value returned is the value from SubjectName with the first letter capitalized. If you need to override this somehow, you can set a different value here.

Tuple Fields of CapitalizeSubjectName

0: String

PossessiveName(String)

PossessiveName(<string>)

When interpolating descriptions, it is often necessary to refer to the creature by name. The interpolation variable poss is used when the creature is referred to as posessing something. By default, the value returned is the value from SubjectName with an apostrophe ‘s’ appended. If you need to override this somehow, you can set a different value here.

Tuple Fields of PossessiveName

0: String

CapitalizePossessiveName(String)

PossessiveName(<string>)

When interpolating descriptions, it is often necessary to refer to the creature by name. The interpolation variable Poss is used when the creature is referred to as posessing something, and this occurs at the beginning of a setence. By default, the value returned is the value from PossessiveName with the first letter capitalized. If you need to override this somehow, you can set a different value here.

Tuple Fields of CapitalizePossessiveName

0: String

Pronoun(String)

Pronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. Various interpolation variables are used to determine these, and all of them can be overridden. You can use this command as a shortcut to overriding all of them based on a linguistic gender. You can pass the following words: “she”, “her”, “he”, “him”, “his”, “they”, “them”, “their”, “xe”, “xem”, “xir”, “it” or “its”. If those words are passed, appropriate values will be overridden for all of the pronoun cases. If any other words are passed to this command, that word will be used in all cases.

Tuple Fields of Pronoun

0: String

SubjectPronoun(String)

SubjectPronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. The interpolation variable subjpro is used when the creature is referred to as the subject of an action. By default, the value returned is ‘it’. If you need to override this somehow, you can set a different value here.

Tuple Fields of SubjectPronoun

0: String

CapitalizeSubjectPronoun(String)

CapitalizeSubjectPronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. The interpolation variable Subjpro is used when the creature is referred to as the subject of an action. By default, the value returned is ‘It’. If you need to override this somehow, you can set a different value here.

Tuple Fields of CapitalizeSubjectPronoun

0: String

PossessivePronoun(String)

PossessivePronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. The interpolation variable posspro is used when the creature is referred to as possessing something. By default, the value returned is ‘its’. If you need to override this somehow, you can set a different value here.

Tuple Fields of PossessivePronoun

0: String

CapitalizePossessivePronoun(String)

CapitalizePossessivePronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. The interpolation variable Posspro is used when the creature is referred to as possessing something. By default, the value returned is ‘Its’. If you need to override this somehow, you can set a different value here.

Tuple Fields of CapitalizePossessivePronoun

0: String

ObjectPronoun(String)

ObjectPronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. The interpolation variable objpro is used when the creature is referred to as the object of an action. By default, the value returned is ‘it’. If you need to override this somehow, you can set a different value here.

Tuple Fields of ObjectPronoun

0: String

ReflexivePronoun(String)

ReflexivePronoun(<string>)

When interpolating descriptions, it is often necessary to use a pronoun for the creature. The interpolation variable refpro is used when the creature is referred to as the object of an action. By default, the value returned is ‘itself’. If you need to override this somehow, you can set a different value here.

Tuple Fields of ReflexivePronoun

0: String

Tiny

Tiny

Sets the size of the creature to tiny, overriding any previous size set. The default size for a creature is medium.

Small

Small

Sets the size of the creature to small, overriding any previous size set. The default size for a creature is medium.

Medium

Medium

Sets the size of the creature to medium, overriding any previous size set. The default size for a creature is medium.

Large

Large

Sets the size of the creature to large, overriding any previous size set. The default size for a creature is medium.

Huge

Huge

Sets the size of the creature to huge, overriding any previous size set. The default size for a creature is medium.

Gargantuan

Gargantuan

Sets the size of the creature to gargantuan, overriding any previous size set. The default size for a creature is medium.

Aberration

Aberration

Sets the type of the creature to “aberration”, overriding any previous type set. The default type for a creature is humanoid.

Beast

Beast

Sets the type of the creature to “beast”, overriding any previous type set. The default type for a creature is humanoid.

Celestial

Celestial

Sets the type of the creature to “celestial”, overriding any previous type set. The default type for a creature is humanoid.

Construct

Construct

Sets the type of the creature to “construct”, overriding any previous type set. The default type for a creature is humanoid.

Dragon

Dragon

Sets the type of the creature to “dragon”, overriding any previous type set. The default type for a creature is humanoid.

Elemental

Elemental

Sets the type of the creature to “elemental”, overriding any previous type set. The default type for a creature is humanoid.

Fey

Fey

Sets the type of the creature to “fey”, overriding any previous type set. The default type for a creature is humanoid.

Fiend

Fiend

Sets the type of the creature to “fiend”, overriding any previous type set. The default type for a creature is humanoid.

Giant

Giant

Sets the type of the creature to “giant”, overriding any previous type set. The default type for a creature is humanoid.

Humanoid

Humanoid

Sets the type of the creature to “humanoid”, overriding any previous type set. The default type for a creature is humanoid.

Monstrosity

Monstrosity

Sets the type of the creature to “monstrosity”, overriding any previous type set. The default type for a creature is humanoid.

Ooze

Ooze

Sets the type of the creature to “ooze”, overriding any previous type set. The default type for a creature is humanoid.

Plant

Plant

Sets the type of the creature to “plant”, overriding any previous type set. The default type for a creature is humanoid.

Undead

Undead

Sets the type of the creature to “undead”, overriding any previous type set. The default type for a creature is humanoid.

CreatureType(String)

CreatureType(<string>)

Sets the type of the creature to some custom value, such as “swarm of Tiny Beasts”, overriding any previous type set. The default type for a creature is humanoid.

Tuple Fields of CreatureType

0: String

Subtype(String)

Subtype(<string>)

Sets the subtype of the creature, such as “goblinoid”, or “dragon”, overriding any previous subtype set. By default, creatures do not have a subtype.

Tuple Fields of Subtype

0: String

Group(String)

Group(<string>)

Sets a field called group on the creature. The usage of this field is unknown, but it might be used to specify that the creature belongs under a different heading (such as demons, oozes, etc.) in a creature listing, that is not related to the subtype.

Tuple Fields of Group

0: String

AnyAlignment

AnyAlignment

Sets the alignment of the creature to “any”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyNonGood

AnyNonGood

Sets the alignment of the creature to “any non-good”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyNonEvil

AnyNonEvil

Sets the alignment of the creature to “any non-evil”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyNonLawful

AnyNonLawful

Sets the alignment of the creature to “any non-lawful”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyNonChaotic

AnyNonChaotic

Sets the alignment of the creature to “any non-chaotic”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyGood

AnyGood

Sets the alignment of the creature to “any good”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyEvil

AnyEvil

Sets the alignment of the creature to “any evil”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyLawful

AnyLawful

Sets the alignment of the creature to “any lawful”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

AnyChaotic

AnyChaotic

Sets the alignment of the creature to “any chaotic”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

LawfulGood

LawfulGood

Sets the alignment of the creature to “lawful good”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

NeutralGood

NeutralGood

Sets the alignment of the creature to “neutral good”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

ChaoticGood

ChaoticGood

Sets the alignment of the creature to “chaotic good”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

LawfulNeutral

LawfulNeutral

Sets the alignment of the creature to “lawful neutral”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

Neutral

Neutral

Sets the alignment of the creature to “neutral”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

ChaoticNeutral

ChaoticNeutral

Sets the alignment of the creature to “chaotic neutral”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

LawfulEvil

LawfulEvil

Sets the alignment of the creature to “lawful evil”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

NeutralEvil

NeutralEvil

Sets the alignment of the creature to “neutral evil”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

ChaoticEvil

ChaoticEvil

Sets the alignment of the creature to “chaotic evil”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

Unaligned

Unaligned

Sets the alignment of the creature to “unaligned”, overriding any previous alignment set. By default, creatures have an alignment of “any”.

Alignment(String)

Alignment(<string>)

Sets the alignment of the creature to some custom value, overriding any previous alignment set. By default, creatures have an alignment of “any”.

Tuple Fields of Alignment

0: String

HitDie(Die)

HitDie(<die-string>)

Sets the base hit die used for calculating hit points on the creature. This only specifies the actual dice used, use ‘HitDiceCount’ to specify the number of dice. The hit points are normally calculated using the die, the count, and a bonus calculated from constitution.

By default, the hit die used is assigned based on the creature’s size, according to the official rules.

Tuple Fields of HitDie

0: Die

HitDiceCount(u8)

HitDiceCount(<integer>)

Sets the number of dice rolled to calculate hit points on the creature. This only specifies the number of cie, use HitDie to override what dice is used. The hit points are normally calculated using the die, the count, and a bonus calculated from constitution. By default, the hit dice count is 1.

Tuple Fields of HitDiceCount

0: u8

HitPoints(u16)

HitPoints(<integer>)

Sets the actual hit points for the creature as shown in the stat block. This does not change the dice expression used for calculation, use HitDie and HitDiceCount for that. By default, the hit points are calculated as the average of the dice expression calculated from those values.

Tuple Fields of HitPoints

0: u16

Armor(Armor)

Armor(<Armor>)

Sets the type of armor worn by the creature, which is used to determine its armor class and the armor class description. See crate::stats::Armor for the available values for this. By default, the creature has no armor.

Tuple Fields of Armor

0: Armor

Shield

Shield

Adds a shield to the creature, which is used to determine its armor class and armor class description.

NoShield

NoShield

Removes a shield from the creature. This is useful if you are overriding a creature that was already specified to have a shield.

Walk(u8)

Walk(<integer>)

Adds a walk speed to the creature. Setting this to 0 will show a speed of 0 ft. By default, the creature’s only speed is a walk speed of 30 ft.

Tuple Fields of Walk

0: u8

Swim(u8)

Swim(<integer>)

Adds a swim speed to the creature. Setting this to 0 will remove the value that was already added. By default, the creature’s only speed is a walk speed of 30 ft.

Tuple Fields of Swim

0: u8

Fly(u8)

Fly(<integer>)

Adds a fly speed to the creature. Setting this to 0 will remove the value that was already added, and remove the hover if that was set. By default, the creature’s only speed is a walk speed of 30 ft.

Tuple Fields of Fly

0: u8

Hover

Hover

Adds a hover tag to the creatures’ flying speed.

Burrow(u8)

Burrow(<integer>)

Adds a burrow speed to the creature. Setting this to 0 will remove the value that was already added. By default, the creature’s only speed is a walk speed of 30 ft.

Tuple Fields of Burrow

0: u8

Climb(u8)

Climb(<integer>)

Adds a climb speed to the creature. Setting this to 0 will remove the value that was already added. By default, the creature’s only speed is a walk speed of 30 ft.

Tuple Fields of Climb

0: u8

SpeedNotes(String)

SpeedNotes(<string>)

Adds custom notes for display in parentheses at the end of the speed property in the final stat-block.

Tuple Fields of SpeedNotes

0: String

Speed(String, u8)

Speed(<string>,<integer>)

Adds a custom speed to the creature with the specified name. Setting this to 0 will remove the value that was already added.

Tuple Fields of Speed

0: String1: u8

Str(u8)

Str(<integer>)

Sets the creature’s Strength score. By default, this score will be 10.

Tuple Fields of Str

0: u8

Dex(u8)

Dex(<integer>)

Sets the creature’s Dexterity score. By default, this score will be 10.

Tuple Fields of Dex

0: u8

Con(u8)

Con(<integer>)

Sets the creature’s Constitution score. By default, this score will be 10.

Tuple Fields of Con

0: u8

Int(u8)

Int(<integer>)

Sets the creature’s Intelligence score. By default, this score will be 10.

Tuple Fields of Int

0: u8

Wis(u8)

Wis(<integer>)

Sets the creature’s Wisdom score. By default, this score will be 10.

Tuple Fields of Wis

0: u8

Cha(u8)

Cha(<integer>)

Sets the creature’s Charisma score. By default, this score will be 10.

Tuple Fields of Cha

0: u8

Saves(Vec<Ability>)

Saves([<Ability>...])

Adds save proficiencies for the specified Abilities. The actual bonuses are calculated from other information. See crate::stats::Ability for possible values.

Tuple Fields of Saves

0: Vec<Ability>

Skills(Vec<Skill>)

Skills([<Skill>...])

Adds skill proficiencies. The bonuses are calculated from other information given. See crate::stats::Skill for possible values.

Tuple Fields of Skills

0: Vec<Skill>

Expertise(Vec<Skill>)

Expertise([<Skill>...])

Adds skill proficiencies with expertise (doubling proficiency bonus for these skills). The bonuses are calculated from other information given. See crate::stats::Skill for possible values.

Tuple Fields of Expertise

0: Vec<Skill>

RemoveSaves(Vec<Ability>)

RemoveSaves([<Ability>...])

Removes save proficiencies for the specified Abilities. See crate::stats::Ability for possible values.

Tuple Fields of RemoveSaves

0: Vec<Ability>

RemoveSkills(Vec<Skill>)

RemoveSkills([<Skill>...])

Removes skill proficiencies and expertise for the specified Skills. See crate::stats::Skill for possible values.

Tuple Fields of RemoveSkills

0: Vec<Skill>

ConditionImmunity(Condition)

ConditionImmunity(<Condition>)

Adds a condition immunity to the creature. See crate::stats::Condition for possible values.

Tuple Fields of ConditionImmunity

0: Condition

Vulnerability(Damage)

Vulnerability(<Damage>)

Adds vulnerability to a damage type. See crate::stats::Damage for possible values.

Tuple Fields of Vulnerability

0: Damage

AllVulnerability

AllVulnerability

Adds vulnerability to all damage.

CustomVulnerability(String)

CustomVulnerability(<string>)

Overrides the text of the vulnerabilities property.

Tuple Fields of CustomVulnerability

0: String

Resistance(Damage)

Resistance(<Damage>)

Adds resistance to a damage type. See crate::stats::Damage for possible values.

Tuple Fields of Resistance

0: Damage

AllResistance

AllResistance

Adds resistance to all damage.

NonmagicalResistance

NonmagicalResistance

Adds resistance to non-magical attacks.

NonSilveredResistance

NonSilveredResistance

Adds resistance to non-silvered and non-magical attacks.

NonAdamantineResistance

NonAdamantineResistance

Adds resistance to non-adamantine and non-magical attacks.

CustomResistance(String)

CustomResistance(<string>)

Overrides the text of the resistances property.

Tuple Fields of CustomResistance

0: String

Immunity(Damage)

Immunity(<Damage>)

Adds immunity to a damage type. See crate::stats::Damage for possible values.

Tuple Fields of Immunity

0: Damage

AllImmunity

AllImmunity

Adds immunity to all damage.

NonmagicalImmunity

NonmagicalImmunity

Adds immunity to non-magical attacks.

NonSilveredImmunity

NonSilveredImmunity

Adds immunity to non-silvered and non-magical attacks.

NonAdamantineImmunity

NonAdamantineImmunity

Adds immunity to non-adamantine and non-magical attacks.

CustomImmunity(String)

CustomImmunity(<string>)

Overrides the text of the immunities property.

Tuple Fields of CustomImmunity

0: String

Languages(Vec<Language>)

Languages([<Language>])

Sets the specified languages to the creature. See crate::stats::Language for possible values.

Tuple Fields of Languages

0: Vec<Language>

UnspokenLanguages(Vec<Language>)

UnspokenLanguages([<Language>])

Adds the specified languages to the creature as unspoken. See crate::stats::Language for possible values.

Tuple Fields of UnspokenLanguages

0: Vec<Language>

Darkvision(u8)

Darkvision(<integer>)

Adds a darkvision sense to the creature with the specified distance range, or removes it if the value is zero.

Tuple Fields of Darkvision

0: u8

Blindsight(u8)

Blindsight(<integer>)

Adds a blindsight sense to the creature with the specified distance range, or removes it if the value is zero.

Tuple Fields of Blindsight

0: u8

BlindsightBlindBeyond(u8)

BlindsightBlindBeyond(<integer>)

Adds a blindsight sense to the creature with the specified distance range and the clause “blind beyond”, or removes it if the value is zero.

Tuple Fields of BlindsightBlindBeyond

0: u8

Truesight(u8)

Truesight(<integer>)

Adds a truesight sense to the creature with the specified distance range, or removes it if the value is zero.

Tuple Fields of Truesight

0: u8

Tremorsense(u8)

Tremorsense(<integer>)

Adds a tremorsense sense to the creature with the specified distance range, or removes it if the value is zero.

Tuple Fields of Tremorsense

0: u8

CustomSense(String, u8)

CustomSense(<string>,<integer>)

Adds a custom sense to the creature with the specified name and distance range, or removes it if the value is zero.

Tuple Fields of CustomSense

0: String1: u8

ExpectNoChallenge

ExpectNoChallenge

“Expects” the creature to have no challenge rating (0 and 0 XP) when the creature is complete.

The challenge rating expectations will cause an error to occur if the calculated challenge rating is different from expected when creature creation is complete. Further expectations will override previous ones. Note that currently the only way to establish a challenge rating is with the Override challenge rating commands. Calculating challenge ratings remains a goal of Monstorr, but has been deferred to some time in the future.

ExpectChallenge(u8)

ExpectChallenge(<integer>)

“Expects” the creature to have a challenge rating of the specified whole number when complete. This number can be 0, which is different from no challenge, as it still provides 10 XP.

The challenge rating expectations will cause an error to occur if the calculated challenge rating is different from expected when creature creation is complete. Further expectations will override previous ones. Note that currently the only way to establish a challenge rating is with the Override challenge rating commands. Calculating challenge ratings remains a goal of Monstorr, but has been deferred to some time in the future.

Tuple Fields of ExpectChallenge

0: u8

ExpectHalfChallenge

ExpectHalfChallenge

“Expects” the creature to have a challenge rating of 1/2 when complete.

The challenge rating expectations will cause an error to occur if the calculated challenge rating is different from expected when creature creation is complete. Further expectations will override previous ones. Note that currently the only way to establish a challenge rating is with the Override challenge rating commands. Calculating challenge ratings remains a goal of Monstorr, but has been deferred to some time in the future.

ExpectQuarterChallenge

ExpectQuarterChallenge

“Expects” the creature to have a challenge rating of 1/4 when complete.

The challenge rating expectations will cause an error to occur if the calculated challenge rating is different from expected when creature creation is complete. Further expectations will override previous ones. Note that currently the only way to establish a challenge rating is with the Override challenge rating commands. Calculating challenge ratings remains a goal of Monstorr, but has been deferred to some time in the future.

ExpectEighthChallenge

ExpectEighthChallenge

“Expects” the creature to have a challenge rating of 1/8 when complete.

The challenge rating expectations will cause an error to occur if the calculated challenge rating is different from expected when creature creation is complete. Further expectations will override previous ones. Note that currently the only way to establish a challenge rating is with the Override challenge rating commands. Calculating challenge ratings remains a goal of Monstorr, but has been deferred to some time in the future.

OverrideNoChallenge

OverrideNoChallenge

Sets the challenge rating to 0 (0 XP).

OverrideChallenge(u8)

OverrideChallenge(<integer>)

Sets the challenge rating to specified whole number. This number can be 0, which is different from no challenge, as it still provides 10 XP.

Tuple Fields of OverrideChallenge

0: u8

OverrideHalfChallenge

OverrideHalfChallenge

Sets the challenge rating to 1/2.

OverrideQuarterChallenge

OverrideQuarterChallenge

Sets the challenge rating to 1/4.

OverrideEighthChallenge

OverrideEighthChallenge

Sets the challenge rating to 1/8.

Multiattack(String, Multiattack)

Multiattack(<string>,<Multiattack>)

Adds a “Multiattack” action to the creature. The first argument is the description to use for the action, which can’t be easily calculated from the other data given. The second argument is a structure which describes what actions are possible in the multiattack, which will be used someday by Monstorr to calculate challenge ratings. Specifying them now will make it easier to get that functionality when it is available.

For more information on the Multiattack argument, see crate::attacks::Multiattack.

Tuple Fields of Multiattack

0: String1: Multiattack

Weapon(Weapon, Option<CompoundAttackEffect>)

Weapon(<Weapon>) Weapon(<Weapon>,optional(<CompoundAttackEffect>))

Adds a built=in weapon attack action to the creature, using default stats to build the description. The optional compound attack effect is for when the weapon causes additional damage or other effects on a hit, due to the monster’s powers, which have nothing to do with the known weapon data. For more information on the arguments, see crate::attacks::Weapon and crate::attacks::CompoundAttackEffect.

Tuple Fields of Weapon

0: Weapon1: Option<CompoundAttackEffect>

ExpectWeaponAttack(Weapon, Attack)

ExpectWeaponAttack(<Weapon>,<Attack>)

Sometimes a small change to another stat can change what a weapon’s attack and hit bonuses, which can change your carefully balanced encounters. If this begins to be a problem, you can use this command to make sure the weapon attack has not changed when you make those small changes. This will report an error if the attack automatically calculated from the weapon is different than what is expected. For information on the arguments, see crate::attacks::Weapon and crate::attacks::Attack.

Tuple Fields of ExpectWeaponAttack

0: Weapon1: Attack

ExpectWeaponEffect(Weapon, AttackEffect)

ExpectWeaponEffect(<Weapon>,<AttackEffect>)

Sometimes a small change to another stat can change what a weapon’s attack and hit bonuses, which can change your carefully balanced encounters. If this begins to be a problem, you can use this command to make sure the weapon’s damage effect has not changed when you make those small changes. This will report an error if the effect automatically calculated from the weapon is different than what is expected. For information on the arguments, see crate::attacks::Weapon and crate::attacks::AttackEffect.

Tuple Fields of ExpectWeaponEffect

0: Weapon1: AttackEffect

OverrideWeaponAttack(Weapon, Attack)

OverrideWeaponAttack(<Weapon>,<Attack>)

If you disagree with the calculated information about a weapon’s attack, you can override it with this command. For information on the arguments, see crate::attacks::Weapon and crate::attacks::Attack.

Tuple Fields of OverrideWeaponAttack

0: Weapon1: Attack

OverrideWeaponEffect(Weapon, AttackEffect)

OverrideWeaponEffect(<Weapon>,<AttackEffect>)

If you disagree with the calculated information about a weapon’s damage effects, you can override it with this command. For information on the arguments, see crate::attacks::Weapon and crate::attacks::Attack.

Tuple Fields of OverrideWeaponEffect

0: Weapon1: AttackEffect

OverrideWeaponDescription(Weapon, String)

OverrideWeaponDescription(<Weapon>,<string>)

If you disagree with the final description created for a weapon’s damage effects, you can override it with this command. This description will be interpolated (see crate::interpolation).

Tuple Fields of OverrideWeaponDescription

0: Weapon1: String

Action(Action, Option<UsageLimit>)

Action(<Action>) Action(<Action>,option(<UsageLimit>))

Adds a non-weapon action to the creature. The description will be calculated based on the information you provide about the action. You can specify a usage limit for the action, such as once per day, requiring a roll on 1d6, etc. For more information on the values you provide here, see crate::actions::Action and crate::actions::UsageLimit.

Tuple Fields of Action

0: Action1: Option<UsageLimit>

OverrideActionDescription(String, String)

OverrideActionDescription(<string>,<string>)

If you disagree with the final description created for an attack, this replaces the calculated description for the action specified in the first argument, with the text in the second. The description will be interpolated (see crate::interpolation).

Tuple Fields of OverrideActionDescription

0: String1: String

RemoveWeapon(Weapon)

RemoveWeapon(<Weapon>)

Removes the first action matching the specified weapon from the list of actions. See crate::attacks::Weapon for possible values.

Tuple Fields of RemoveWeapon

0: Weapon

MoveWeapon(Weapon, i8)

`MoveWeapon(,)

Moves a weapon down or up on the list. This is useful when you are overriding an existing creature and want to ensure that actions appear in the correct order.

Tuple Fields of MoveWeapon

0: Weapon1: i8

RemoveAction(String)

RemoveAction(<string>)

Removes the first action with the specified name from the list of actions.

Tuple Fields of RemoveAction

0: String

MoveAction(String, i8)

`MoveAction(,)

Moves an action down or up on the list. This is useful when you are overriding an existing creature and want to ensure that actions appear in the correct order.

Tuple Fields of MoveAction

0: String1: i8

Reaction(Reaction, Option<UsageLimit>)

Reaction(<Reaction>) Reaction(<Reaction>,option(<UsageLimit>))

Adds a reaction. For more information on the arguments, see crate::reactions::Reaction and crate::actions::UsageLimit.

Tuple Fields of Reaction

0: Reaction1: Option<UsageLimit>

RemoveReaction(String)

RemoveReaction(<string>)

Removes the first reaction with the specified name.

Tuple Fields of RemoveReaction

0: String

Feature(Feature, Option<UsageLimit>)

Feature(<Feature>) Feature(<Feature>,option(<UsageLimit>))

Adds a feature. Some features have effects on other parts of the stat block. Do not use this to add spellcasting and innate spellcasting features, those have their own commands. For more information on the arguments, see crate::features::Feature and crate::actions::UsageLimit.

Tuple Fields of Feature

0: Feature1: Option<UsageLimit>

RemoveFeature(String)

RemoveFeature(<string>)

Removes the first feature with the specified name. This can also be used to remove spellcasting and innate spellcasting.

Tuple Fields of RemoveFeature

0: String

Spellcasting(Vec<SpellcastingCommand>)

Spellcasting([<SpellcastingCommand>...])

Calls the listed commands to add spellcasting features to the creature. For more information on what commands are available, see SpellcastingCommand.

Tuple Fields of Spellcasting

0: Vec<SpellcastingCommand>

InnateSpellcasting(Vec<InnateSpellcastingCommand>)

InnateSpellcasting([<InnateSpellcastingCommand>...])

Calls the listed commands to add innate spellcasting features to the creature. For more information on what commands are available, see InnateSpellcastingCommand.

Tuple Fields of InnateSpellcasting

0: Vec<InnateSpellcastingCommand>

LegendaryActions(u8, Vec<LegendaryAction>)

LegendaryActions(<integer>,[<LegendaryAction>...])

Adds legendary actions to the creature. The first argument is the number of legendary actions the creature can take per round. The second is the list of actions. For more information on this, see crate::actions::LegendaryAction.

Tuple Fields of LegendaryActions

0: u81: Vec<LegendaryAction>

RemoveLegendaryAction(String)

RemoveLegendaryAction(<string>)

Removes the specified legendary action by name from the creature. When the last one is removed, will remove the legendary action feature.

Tuple Fields of RemoveLegendaryAction

0: String

LairActions(String, Vec<String>, String)

Currently Unsupported

Tuple Fields of LairActions

0: String1: Vec<String>2: String

RegionalEffects(String, Vec<String>, String)

Currently Unsupported

Tuple Fields of RegionalEffects

0: String1: Vec<String>2: String

Implementations

impl CreatureCommand

fn execute(
&self,
working_dir: &PathBuf,
creature: &mut Creature,
hooks: &mut CreatureCreationHooks
) -> Result<(), CreatureError>