Module monstorr_lib::interpolation

In Monstorr, interpolation is the process of substituting variables and expressions with actual values at the last minute. There are actually two types of interpolation that happen in Monstorr, one occurs during the generation of stat blocks from feature descriptions and other strings, the other occurs when including an external creature file into another (see crate::creature_commands::CreatureCommand::Include). They both work basically the same, but have slightly different inputs and outputs.

Strings which can be interpolated contain normal text interspersed with interpolation expressions delimited from the rest. In order to make it easier to include feature interpolation strings inside an included document, the delimiters for feature interpolation and inclusion interpolation are different.

The result of an interpolation differs between the two modes. For inclusion, the result is a plain string that will then be deserialized into creature commands. For descriptions, however, the result is structured text.

Expression Delimiters

When embedding interpolation in the included file, delimit expressions with ‘$<…>’. When embedding interpolation in features, delimit the expressions with ‘${…}’.

All text outside these delimiters are treated as normal text. While you can use the backslash (‘') to escape the dollar sign character (’$’), and other backslashes, this is not necessary unless, for some reason, you need to include the ‘${’ character combination in regular text, or worse, the ‘${’ character combination. In the former case you can produce that with ‘${’. In the latter, you can produce it with ‘\${’. Outside of those instances, a dollar sign and a backslash are treated as normal characters. In interpolating for inclusion, the same applies to ‘$<’ and ‘$<’ character combinations.

Expression Syntax

The expression syntax is fairly simple. The following values operations are supported:

Variable References. A plain identifier is a variable reference. Variables are provided by Monstorr according to the mode of interpolation. In inclusion, they are string values provided as arguments by the Include command, see crate::creature_commands::CreatureCommand::Include. In descriptions, the variables are properties of the creature, which may be strings, numbers or dice. For available creature properties, see crate::creature::Creature. If the variable does not exist, or returns no value, an error will occur when it is referenced.

Strings. Strings are delimited by double-quotes. They return a string value.

Numbers. Only integers are supported, and consist of a series of decimal digits. They return a number value.

Dice. Dice literals are supported in forms like “1d8” and “2d6”. Both numbers are required. They return a dice expression value. When a dice expression is converted to a string, it is formatted as in standard descriptions with the average followed by the actual expression in parentheses.

Parenthesis. Expressions in between parentheses ‘(..)’ will be calculated first.

Properties and Index. Some values returned by variables are objects that have properties, or lists that have indexes. Properties are accessed by using the ‘.’ operator followed by the property’s identifier. List items are accessed by using dot operator followed by a number indicating the index to access.

Negating Numbers. The ‘-’ can be used before a number to negate it. The value must be a number, no other value types can be negated. Remember that if you negate a value that occurs after another expression, it may be confused with the minus operator. Use parentheses or replace it with ‘+ -’ to remove this ambiguity.

Signing Numbers and Dice. Numbers and dice expressions can be signed, so that when they are formatted as strings, a ‘+’ appears before them if they are positive. The ‘-’ will always appear if they are negative. The ‘+’ operator which appears before the value will ensure that that takes place. It is infectious, so if you add other values to the expression, the final result of the expression will be formatted that way. Remember that if you sign a value that occurs after another expression, it may be confused with the plus operator. Use parentheses or replace it with ‘+ +’ to remove this ambiguity.

Stringifying Values. Numbers and dice are automatically converted to strings when they are the final result of an expression. However, for type-safety, they are not if they are used in a more compound expression, and concatenating strings and non-strings is not allowed. This prevents you from making mistakes like ‘“foo” + 1 + 2’ when what you really meant was ‘“foo” (+1 + 2)’. The ‘$’ operator can be used before a number or dice value to convert it to a string, making concatenation possible.

Multiplication. To multiply two number expressions, use the ‘*’ operator. You can multiply dice by numbers to get a new dice expression. If the number is not 1 or -1, then your dice expression will be formatted with a factor. You can not multiply two dice values together, as the result is undetermined. Remember, the coefficient of a dice value is not the same as a numeric factor: “2d6” is not the same as “1d6 x 2”, as they have different sets of possible results.

Division. As only integer number values are allowed, it is necessary to know how to deal with fractions when dividing them. For that reason, there are two division operators. Use “/<” to divide the numbers and round the result down, use “/>” to divide the numbers and round the result up. Dice values can be divided by numbers, which will result in them having a factor expression when formatted. However, nothing can be divided by dice expressions, as the result of that is undetermined.

Addition, Subtraction and Concatenation. To add and subtract number and dice values, use the “+” and “-” characters respectively. Adding dice together can create more complex dice expressions which involve several dice terms. If you add the same dice together, it should result in simply changing the coefficient. For example, “2d6 + 1d6” should result in “3d6”. However, “2d4 - 1d4” would not be the same as “1d4”, as they would have different sets of results (the first would have values ranging from -1 to 7, while the other has values ranging from 1-4, and that doesn’t even consider distribution )

Consecutive Expressions. If two expressions are included consecutive to each other, then the results are both pushed out to the resulting string in order. For example, ‘“foo” “bar”’ would output the text “foobar”.

Structured Text Functions. The result of interpolating a description is not just a simple string, it’s a list of text blocks that can contain styled text. The actual result is a list of crate::structured_text::TextBlocks. There are two kinds of text blocks, and each of these can contain lists of four different kinds of spans. These are achieved through special functions.

A structured text function consists of a keyword, followed by an optional parentheses expression. Unlike normal expressions, these parentheses can contain normal template text, thus “italic( } italicized text ${ )” can be done. There are four such functions, which modify the structured text output.

par: starts a new normal paragraph block, allowing you to break up your description text. There is a variant of this function which takes an argument in parentheses, but it should not be used, as it has the result of creating what looks like a new feature in the stat block.

sub(...), sub: starts a subparagraph, with a heading contained in the parenthesis. In official content, a subparagraph in this case would create a paragraph with a hanging indent. The heading is an optional emphasized text that will appear at the beginning of the first line. In the official content, subparagraphs are used in the metallic dragon breath weapons, and in spellcasting spell lists.

‘italic(…)’: sets the text enclosed in the parentheses as italic. Nesting italic text inside italic text is not allowed, as it has undetermined effects. You can nest bold text inside italics, however.

‘bold(…)’: sets the text enclosed in the parentheses as bold. Nesting bold text inside bold text is not allowed, as it has undetermined effects. You can nest italic text inside bold, however.

Examples

For examples of interpolation, look at the built-in creatures, whose raw files can be retrieved using the validate command from the Monstorr command line.

Structs

Document

InterpolationTokenizer

OverriddenInterpolationObject

TokenDetails

Enums

InterpolationMode

InterpolationOperation

InterpolationState

InterpolationValue

StructureKeyword

Traits

InterpolationObject

Functions

interpolate_str_for_deserialization

interpolate_str_for_statblock