Module monstorr_lib::stat_block

Creature stat blocks are a simple property/value mapping, with no functionality. They are intended to be output directly as JSON, or used in a template to generate content for publication.

All property values are strings or lists of strings. Numbers and other data have been transformed into the standard formatting used in the core books. For example, abilities show their score and the modifier in parentheses, skills are already joined into a list with their bonus numbers, etc. Most templates should be able to simply output the property values as-is. The only complexity lay in the descriptions, which contain structured text.

The only exception is the ‘languages’ property, which does not display an em-dash if it is not present.

The schema for the JSON format is defined on the CreatureStatBlock struct.

Structs

CreatureStatBlock

This is the final object output to JSON, or passed through the template at the end. The properties are listed with their types. If a property is optional, it is marked with a ‘?’.

StatBlockFeature

Features such as special abilities and actions in a stat block are a simple object that contains a single ‘text’ property list of crate::structured_text::TextBlock. The added property makes JSON formatting of the structure a little less confusing visually, and makes it easier to wrap the object in template code.

StatBlockLairActions

StatBlockLegendary

StatBlockRegionalEffects

Traits

TryIntoStatBlock