Struct monstorr_lib::stat_block::CreatureStatBlock

pub struct CreatureStatBlock {

    pub name: String,
    pub size: String,
    pub type_: String,
    pub subtype: Option<String>,
    pub group: Option<String>,
    pub alignment: String,
    pub armor: String,
    pub hit_points: String,
    pub speed: String,
    pub strength: String,
    pub dexterity: String,
    pub constitution: String,
    pub intelligence: String,
    pub wisdom: String,
    pub charisma: String,
    pub saving_throws: Option<String>,
    pub skills: Option<String>,
    pub damage_vulnerabilities: Option<String>,
    pub damage_resistances: Option<String>,
    pub damage_immunities: Option<String>,
    pub condition_immunities: Option<String>,
    pub senses: String,
    pub languages: Option<String>,
    pub challenge_rating: String,
    pub special_abilities: Vec<StatBlockFeature>,
    pub actions: Vec<StatBlockFeature>,
    pub reactions: Vec<StatBlockFeature>,
    pub legendary_actions: Option<StatBlockLegendary>,
    pub lair_actions: Option<StatBlockLairActions>,
    pub regional_effects: Option<StatBlockRegionalEffects>,
    pub source: Option<String>,

}

This is the final object output to JSON, or passed through the template at the end. The properties are listed with their types. If a property is optional, it is marked with a ‘?’.

Fields
name: String

name: <string>

size: String

size: <string>

type_: String

type: <string>

subtype: Option<String>

subtype?: <string>

group: Option<String>

group?: <string>

alignment: String

alignment: <string>

armor: String

armor: <string>

hit_points: String

hit_points: <string>

speed: String

speed: <string>

strength: String

strength: <string>

dexterity: String

dexterity: <string>

constitution: String

constitution: <string>

intelligence: String

intelligence: <string>

wisdom: String

wisdom: <string>

charisma: String

charisma: <string>

saving_throws: Option<String>

saving_throws?: <string>

skills: Option<String>

skills?: <string>

damage_vulnerabilities: Option<String>

damage_vulnerabilities?: <string>

damage_resistances: Option<String>

damage_resistances?: <string>

damage_immunities: Option<String>

damage_immunities?: <string>

condition_immunities: Option<String>

condition_immunities?: <string>

senses: String

senses: <string>

languages: Option<String>

languages?: <string>

If there are no languages, then this property is empty. In standard formatting, this is often replaced with an em dash if there are no languages. I find this inconsistent with the other properties (the same is not done for skills and saves), and leave that decision up to the creator.

challenge_rating: String

challenge_rating: <string>

special_abilities: Vec<StatBlockFeature>

special_abilities?: list(<StatBlockFeature>)

actions: Vec<StatBlockFeature>

actions?: list(<StatBlockFeature>)

reactions: Vec<StatBlockFeature>

reactions?: list(<StatBlockFeature>)

legendary_actions: Option<StatBlockLegendary>

legendary_actions?: <StatBlockLegendary>

lair_actions: Option<StatBlockLairActions>

lair_actions?: <StatBlockLairActions>

Lair actions are not yet supported.

regional_effects: Option<StatBlockRegionalEffects>

regional_effects: <StatBlockRegionalEffects>

Regional effects are not yet supported.

source: Option<String>

source?: <string>

Implementations

impl CreatureStatBlock

pub fn write_to_string(&self, ugly: bool) -> Result<String, String>