Module monstorr_lib::stats

Structs

ParseAbilityError

ParseDamageError

ParseLanguageError

Enums

Ability

These are used in a few creature commands and elsewhere to indicate an ability score.

Alignment

Armor

Represents the possible types of armor from the player’s handbook, plus a couple of others.

ChallengeRating

Condition

CreatureSize

CreatureType

Damage

These values are used to represent standard damage types.

DamageRestriction

Language

These values are used to represent most of the standard languages found in the core rulebooks, as well as custom languages.

ParseChallengeRatingError

Skill

These are used to add skill proficiencies to a creature.