Struct monstorr_lib::creature::Creature

pub struct Creature {

    pub name: String,
    pub subject: Option<String>,
    pub subject_cap: Option<String>,
    pub possessive: Option<String>,
    pub possessive_cap: Option<String>,
    pub size: CreatureSize,
    pub type_: CreatureType,
    pub subtype: Option<String>,
    pub group: Option<String>,
    pub alignment: Alignment,
    pub hit_die: Die,
    pub hit_dice_count: u8,
    pub override_hit_points: Option<u16>,
    pub armor: CreatureArmor,
    pub speed: CreatureSpeed,
    pub strength: u8,
    pub dexterity: u8,
    pub constitution: u8,
    pub intelligence: u8,
    pub wisdom: u8,
    pub charisma: u8,
    pub strength_save: bool,
    pub dexterity_save: bool,
    pub constitution_save: bool,
    pub intelligence_save: bool,
    pub wisdom_save: bool,
    pub charisma_save: bool,
    pub skills: CreatureProficiencies,
    pub vulnerabilities: CreatureResistances,
    pub resistances: CreatureResistances,
    pub immunities: CreatureResistances,
    pub condition_immunities: CreatureConditionImmunities,
    pub senses: CreatureSenses,
    pub languages: Vec<(Language, bool)>,
    pub challenge_rating: ChallengeRating,
    pub multiattack: Option<Multiattack>,
    pub special_abilities: Vec<CreatureSpecialAbility>,
    pub actions: Vec<CreatureAction>,
    pub reactions: Vec<CreatureFeature>,
    pub legendary_actions: Option<CreatureLegendaryActions>,
    pub source: Option<String>,

}

The following properties can be retrieved during interpolation. Their value types are listed in parentheses after.

name (string): The name of the creature

subj (string): The name of the creature as used for the subject of an action.

Subj (string): The subject of the creature capitalized.

poss (string): The name of the creature in possessive form.

Poss (string): The capitalized possessive.

size (string): The size of the creature

type (string): The type of the creature

subtype (string): The subtype of the creature, or an empty string

group (string): The creature’s group, or an empty string.

alignment (string): The alignment of the creature

hit_dice (dice): The calculated hit dice for the creature.

hit_points (number): The calculate hit points for the creature.

armor_class (number): The calculated armor class for the creature.

strength (number): The strength score for the creature.

dexterity (number): The dexterity score for the creature.

constitution (number): The constitution score for the creature.

intelligence (number): The intelligence score for the creature.

wisdom (number): The wisdom score for the creature.

charisma (number): The charisma score for the creature.

atk (number): The best of the dexterity and strength modifiers.

spell_atk (number): The calculated spell attack bonus. If the creature has both innate and regular spellcasting, this will return the innate bonus. If the creature has neither, this property will not be available, and an error will occur.

str (number): The calculated strength modifier.

dex (number): The calculated dexterity modifier.

con (number): The calculated constitution modifier.

int (number): The calculated intelligence modifier.

wis (number): The calculated wisdom modifier.

cha (number): The calculated charisma modifier.

str_save (number): The calculated save bonus for strength.

dex_save (number): The calculated save bonus for dexterity.

con_save (number): The calculated save bonus for constitution.

int_save (number): The calculated save bonus for intelligence.

wis_save (number): The calculated save bonus for wisdom.

cha_save (number): The calculated save bonus for charisma.

prof (number): The proficiency bonus, taken from the creatures challenge rating.

Fields
name: Stringsubject: Option<String>subject_cap: Option<String>possessive: Option<String>possessive_cap: Option<String>size: CreatureSizetype_: CreatureTypesubtype: Option<String>group: Option<String>alignment: Alignmenthit_die: Diehit_dice_count: u8override_hit_points: Option<u16>armor: CreatureArmorspeed: CreatureSpeedstrength: u8dexterity: u8constitution: u8intelligence: u8wisdom: u8charisma: u8strength_save: booldexterity_save: boolconstitution_save: boolintelligence_save: boolwisdom_save: boolcharisma_save: boolskills: CreatureProficienciesvulnerabilities: CreatureResistancesresistances: CreatureResistancesimmunities: CreatureResistancescondition_immunities: CreatureConditionImmunitiessenses: CreatureSenseslanguages: Vec<(Language, bool)>challenge_rating: ChallengeRatingmultiattack: Option<Multiattack>special_abilities: Vec<CreatureSpecialAbility>actions: Vec<CreatureAction>reactions: Vec<CreatureFeature>legendary_actions: Option<CreatureLegendaryActions>source: Option<String>
Implementations

impl Creature

fn get_subject(&self, capitalize: bool) -> String

fn get_possessive(&self, capitalize: bool) -> String

fn get_hit_dice(&self) -> DiceExpression

fn saving_throw_to_stat_block(ability: Ability, creature: &Creature) -> i8

fn saving_throws_to_stat_block(creature: &Creature) -> Option<String>

fn languages_to_stat_block(list: &Vec<(Language, bool)>) -> Option<String>

fn actions_to_stat_block(
vec: &Vec<CreatureAction>,
data: &Rc<Creature>
) -> Result<Vec<StatBlockFeature>, InterpolationErrorDetails>

fn features_to_stat_block(
vec: &Vec<CreatureFeature>,
data: &Rc<Creature>
) -> Result<Vec<StatBlockFeature>, InterpolationErrorDetails>

fn spellcasting_to_stat_block(
spells: &Spellcasting,
data: &Rc<Creature>
) -> Result<StatBlockFeature, InterpolationErrorDetails>

fn innate_spellcasting_to_stat_block(
spells: &InnateSpellcasting,
data: &Rc<Creature>
) -> Result<StatBlockFeature, InterpolationErrorDetails>

fn special_abilities_to_stat_block(
vec: &Vec<CreatureSpecialAbility>,
data: &Rc<Creature>
) -> Result<Vec<StatBlockFeature>, InterpolationErrorDetails>

fn get_armor_class(&self) -> u8

fn armor_to_stat_block(&self) -> String

pub fn set_source(&mut self, name: &str)

pub fn set_name(&mut self, name: &str)

pub fn set_subject(&mut self, name: &str)

pub fn set_capitalized_subject(&mut self, name: &str)

pub fn set_possessive(&mut self, name: &str)

pub fn set_capitalized_possessive(&mut self, name: &str)

pub fn set_tiny(&mut self)

pub fn set_small(&mut self)

pub fn set_medium(&mut self)

pub fn set_large(&mut self)

pub fn set_huge(&mut self)

pub fn set_gargantuan(&mut self)

pub fn set_aberration(&mut self)

pub fn set_beast(&mut self)

pub fn set_celestial(&mut self)

pub fn set_construct(&mut self)

pub fn set_dragon(&mut self)

pub fn set_elemental(&mut self)

pub fn set_fey(&mut self)

pub fn set_fiend(&mut self)

pub fn set_giant(&mut self)

pub fn set_humanoid(&mut self)

pub fn set_monstrosity(&mut self)

pub fn set_ooze(&mut self)

pub fn set_plant(&mut self)

pub fn set_undead(&mut self)

pub fn set_custom_type(&mut self, name: &str)

pub fn set_subtype(&mut self, name: &str)

pub fn set_group(&mut self, group: &str)

pub fn set_any_alignment(&mut self)

pub fn set_any_non_good(&mut self)

pub fn set_any_non_evil(&mut self)

pub fn set_any_non_lawful(&mut self)

pub fn set_any_non_chaotic(&mut self)

pub fn set_any_good(&mut self)

pub fn set_any_evil(&mut self)

pub fn set_any_lawful(&mut self)

pub fn set_any_chaotic(&mut self)

pub fn set_lawful_good(&mut self)

pub fn set_neutral_good(&mut self)

pub fn set_chaotic_good(&mut self)

pub fn set_lawful_neutral(&mut self)

pub fn set_neutral(&mut self)

pub fn set_chaotic_neutral(&mut self)

pub fn set_lawful_evil(&mut self)

pub fn set_neutral_evil(&mut self)

pub fn set_chaotic_evil(&mut self)

pub fn set_unaligned(&mut self)

pub fn set_custom_alignment(&mut self, name: &str)

pub fn set_hit_die(&mut self, die: &Die)

pub fn set_hit_dice_count(&mut self, count: &u8)

pub fn set_hit_points_override(&mut self, points: &u16)

pub fn set_armor(&mut self, armor: &Armor)

pub fn enable_shield(&mut self)

pub fn disable_shield(&mut self)

pub fn walk(&mut self, speed: &u8)

pub fn swim(&mut self, speed: &u8)

pub fn fly(&mut self, speed: &u8)

pub fn burrow(&mut self, speed: &u8)

pub fn climb(&mut self, speed: &u8)

pub fn enable_hover(&mut self)

pub fn speed_notes(&mut self, notes: &str)

pub fn custom_speed(&mut self, movement: &str, speed: &u8)

pub fn set_str(&mut self, score: &u8)

pub fn set_dex(&mut self, score: &u8)

pub fn set_con(&mut self, score: &u8)

pub fn set_int(&mut self, score: &u8)

pub fn set_wis(&mut self, score: &u8)

pub fn set_cha(&mut self, score: &u8)

pub fn add_saves(&mut self, abilities: &[Ability])

pub fn add_skills(&mut self, skills: &[Skill])

pub fn add_expertise(&mut self, skills: &[Skill])

pub fn remove_saves(&mut self, abilities: &[Ability])

pub fn remove_skills(&mut self, skills: &[Skill])

pub fn add_condition_immunity(&mut self, condition: &Condition)

pub fn add_vulnerability(&mut self, damage: &Damage)

pub fn add_all_vulnerability(&mut self)

pub fn add_custom_vulnerability(&mut self, name: &str)

pub fn add_resistance(&mut self, damage: &Damage)

pub fn add_all_resistance(&mut self)

pub fn add_custom_resistance(&mut self, name: &str)

pub fn add_nonmagical_resistance(&mut self)

pub fn add_nonsilvered_resistance(&mut self)

pub fn add_nonadamantine_resistance(&mut self)

pub fn add_immunity(&mut self, damage: &Damage)

pub fn add_all_immunities(&mut self)

pub fn add_custom_immunity(&mut self, name: &str)

pub fn add_nonmagical_immunity(&mut self)

pub fn add_nonsilvered_immunity(&mut self)

pub fn add_nonadamantine_immunity(&mut self)

pub fn add_languages(&mut self, languages: &[Language])

pub fn add_unspoken_languages(&mut self, languages: &[Language])

pub fn add_darkvision(&mut self, distance: &u8)

pub fn add_blindsight(&mut self, distance: &u8)

pub fn add_blindsight_blind_beyond(&mut self, distance: &u8)

pub fn add_truesight(&mut self, distance: &u8)

pub fn add_tremorsense(&mut self, distance: &u8)

pub fn add_custom_sense(&mut self, sense: &str, distance: &u8)

pub fn set_no_challenge_rating(&mut self)

pub fn set_challenge_rating(&mut self, cr: &u8)

pub fn set_half_challenge_rating(&mut self)

pub fn set_quarter_challenge_rating(&mut self)

pub fn set_eighth_challenge_rating(&mut self)

fn find_action(&self, name: &str) -> Option<&CreatureAction>

fn find_weapon(&self, weapon: &Weapon) -> Option<&CreatureAction>

pub fn find_action_mut(&mut self, name: &str) -> Option<&mut CreatureAction>

fn find_weapon_mut(&mut self, weapon: &Weapon) -> Option<&mut CreatureAction>

pub fn check_multiattack(
&self,
details: &Multiattack
) -> Result<(), CreatureError>

pub fn set_multiattack(&mut self, description: String, details: &Multiattack)

pub fn add_weapon(
&mut self,
weapon: &Weapon,
compound: &Option<CompoundAttackEffect>
)

pub fn expect_weapon_attack(
&self,
weapon: &Weapon,
attack: &Attack
) -> Result<(), CreatureError>

pub fn expect_weapon_effect(
&self,
weapon: &Weapon,
effect: &AttackEffect
) -> Result<(), CreatureError>

pub fn override_weapon_attack(
&mut self,
weapon: &Weapon,
attack: Attack
) -> Result<(), CreatureError>

pub fn override_weapon_effect(
&mut self,
weapon: &Weapon,
effect: AttackEffect
) -> Result<(), CreatureError>

pub fn override_weapon_description(
&mut self,
weapon: &Weapon,
description: String
) -> Result<(), CreatureError>

pub fn add_action(&mut self, action: &Action, usage_limit: &Option<UsageLimit>)

pub fn override_action_description(
&mut self,
name: &str,
description: String
) -> Result<(), CreatureError>

pub fn remove_weapon(&mut self, weapon: &Weapon)

pub fn remove_action(&mut self, name: &str)

pub fn add_reaction(
&mut self,
reaction: Reaction,
usage_limit: Option<UsageLimit>
)

pub fn remove_reaction(&mut self, name: &str)

pub fn add_feature(&mut self, feature: Feature, usage_limit: Option<UsageLimit>)

pub fn remove_feature(&mut self, name: &str)

pub fn remove_legendary_action(&mut self, name: &str)

pub fn get_spellcasting(&self) -> Option<&Spellcasting>

pub fn get_or_add_spellcasting_mut(&mut self) -> &mut Spellcasting

pub fn get_innate_spellcasting(&self) -> Option<&InnateSpellcasting>

pub fn get_or_add_innate_spellcasting_mut(&mut self) -> &mut InnateSpellcasting

pub fn get_spell_attack_bonus(&self) -> Option<i8>

pub fn set_legendary_actions(
&mut self,
description: String,
actions: Vec<CreatureLegendaryAction>
)