Enum monstorr_lib::features::Feature

pub enum Feature {

    Aggressive,
    Ambusher,
    Amorphous,
    Amphibious,
    AngelicWeapons(DiceExpression, Vec<String>),
    AntimagicSusceptibility,
    BlindSenses,
    BloodFrenzy,
    Brute(Vec<String>),
    Charge(u8, String, DiceExpression, Damage),
    DamageTransfer,
    DeathBurst(String, AttackEffect, Option<CompoundAttackEffect>),
    DevilSight,
    Echolocation,
    FalseAppearance(String),
    FeyAncestry,
    Flyby,
    Grappler,
    HoldBreath(u8, bool),
    Illumination(u8),
    ImmutableForm,
    IncorporealMovement(DiceExpression),
    Inscrutable,
    Invisibility,
    KeenSenses(bool, bool, bool),
    LabyrinthineRecall,
    LegendaryResistance,
    LightSensitivity,
    MagicResistance,
    MagicWeapons,
    MartialAdvantage(DiceExpression),
    Mimicry(String, u8),
    NimbleEscape,
    PackTactics,
    Pounce(u8, u8),
    Rampage,
    Reactive,
    Reckless,
    Regeneration(u8, Vec<Damage>),
    Rejuvenation(String),
    Relentless(u8),
    ShadowStealth,
    Shapechanger(String),
    SiegeMonster,
    SpiderClimb,
    StandingLeap(u8),
    Steadfast,
    Stench(u8, u8),
    SunlightSensitivity,
    SureFooted,
    SurpriseAttack(DiceExpression),
    Swarm,
    Tunneler(u8),
    TurnResistance,
    TwoHeads,
    UndeadFortitude,
    WebSense,
    WebWalker,
    Feature(String, String),

}

Many of the variants in this value are attempts to mimic common features found in the SRD, as well as actions which the Dungeon Master’s Guide says may have an effect on the challenge rating. Other features make automatic changes to the creature at the end of creation. However, since the SRD does not use consistent phrasing for all features, you may not get the output you expect. If you want to use a feature with one of the specified names, use that, even if you have to override the description to fix it later. That way, you’ll get all of the benefits of the feature.

Features which make changes to the creature are done after the creature is mostly finished. Therefore it is not necessary to place the feature after other required features.

Variants

Aggressive

Aggressive

Ambusher

Ambusher

Amorphous

Amorphous

Amphibious

Amphibious

AngelicWeapons(DiceExpression, Vec<String>)

AngelicWeapons(<dice-expression-string>,[<string>..])

The first argument is the additional damage added to weapon attacks. The second is the list of actions this is automatically applied to.

This feature will automatically change the specified attacks to add the specified damage. It will cause an error if the actions do not exist, or if they are in a configuration it can not deal with.

AntimagicSusceptibility

AntimagicSusceptibility

BlindSenses

BlindSenses

BloodFrenzy

BloodFrenzy

Brute(Vec<String>)

Brute([<string>...])

The argument is the list of melee attacks this feature is applied to.

This feature automatically makes changes to the specified actions. An error will occur if the actions do not exist, are not melee actions, or are in a configuration it does not know how to deal with.

Charge(u8, String, DiceExpression, Damage)

Charge(<integer>,<string>,<dice-expression-string>,<Damage>)

The first argument is the distance of the charge, the second is the type of attack, the third and fourth indicate the extra damage the charge can cause.

DamageTransfer

DamageTransfer

DeathBurst(String, AttackEffect, Option<CompoundAttackEffect>)

DeathBurst(<string>,<AttackEffect>) DeathBurst(<string>,<AttackEffect>,option(<CompoundAttackEffect>))

The first is the description of the explosion, the rest is the effect of the explosion.

DevilSight

DevilSight

Echolocation

Echolocation

FalseAppearance(String)

FalseAppearance(<string>)

The argument is a description of the false appearance form.

FeyAncestry

FeyAncestry

Flyby

Flyby

Grappler

Grappler

HoldBreath(u8, bool)

HoldBreath(<integer>,<boolean>)

The first is the duration, the second is whether this is reversed for aquatic creatures (i.e. the creature must hold its breath out of water)

Illumination(u8)

Illumination(<integer>)

The integer is the radius of the illumination.

ImmutableForm

ImmutableForm

IncorporealMovement(DiceExpression)

IncorporealMovement(<dice-expression-string>)

The argument specifies the force damage roll the creature takes if it ends its turn inside an object.

Inscrutable

Inscrutable

Invisibility

Invisibility

KeenSenses(bool, bool, bool)

KeenSenses(<boolean>,<boolean>,<boolean>)

The arguments specify which senses this applies. The first specifies sight, the second hearing, and the third smell.

LabyrinthineRecall

LabyrinthineRecall

LegendaryResistance

LegendaryResistance

LightSensitivity

LightSensitivity

MagicResistance

MagicResistance

MagicWeapons

MagicWeapons

MartialAdvantage(DiceExpression)

MartialAdvantage(<dice-expression-string>)

The argument specifies the extra damage dice the advantage gives them.

Mimicry(String, u8)

Mimicry(<string>,<integer>)

The first argument describes the mimicked speech, the second is the DC for an insight check.

NimbleEscape

NimbleEscape

PackTactics

PackTactics

Pounce(u8, u8)

Pounce(<integer>,<integer>)

The first argument is the distance of the pounce, the second is the Strength save DC.

Rampage

Rampage

Reactive

Reactive

Reckless

Reckless

Regeneration(u8, Vec<Damage>)

Regeneration(<integer>,[<Damage>...])

The first argument is the number of hit points regenerated, the second is the list of damage this doesn’t apply to.

Rejuvenation(String)

Rejuvenation(<string>)

The argument is the description of the feature, as every one is different.

Relentless(u8)

Relentless(<integer>)

The argument specifies the damage limit.

ShadowStealth

ShadowStealth

Shapechanger(String)

Shapechanger(<string>)

The argument describes the shapes available.

SiegeMonster

SiegeMonster

SpiderClimb

SpiderClimb

StandingLeap(u8)

StandingLeap(<integer>)

The argument specifies the distance of the leap.

Steadfast

Steadfast

Stench(u8, u8)

Stench(<integer>,<integer>)

The first integer specifies the range, the second the Constitution save DC.

SunlightSensitivity

SunlightSensitivity

SureFooted

SureFooted

SurpriseAttack(DiceExpression)

SurpriseAttack(<dice-expression-string>)

The argument specifies the extra damage incurred.

Swarm

Swarm

Tunneler(u8)

Tunneler(<integer>)

The argument specifies the width of the tunnel.

TurnResistance

TurnResistance

TwoHeads

TwoHeads

UndeadFortitude

UndeadFortitude

WebSense

WebSense

WebWalker

WebWalker

Feature(String, String)

Feature(<string>,<string>)

Adds a custom feature to the creature, with the specified name and description. These will be interpolated.

Implementations

impl Feature

pub fn get_name(&self) -> String

pub fn get_description(&self) -> String